experiments/pi5-vlm-test/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.45.0
accelerate>=0.25.0
pillow>=10.0.0
numpy>=1.24.0
qwen-vl-utils
opencv-python-headless>=4.8.0
einops>=0.7.0
timm>=0.9.0
sentencepiece>=0.1.99
